Dungeons-Droids-main/requirements.txt
flask
openai
python-dotenv
pymongo
bcrypt
